requirements.txt
-i https://pypi.org/simple
asgiref==3.2.7
boto3==1.13.12
botocore==1.16.12
dj-database-url==0.5.0
django-heroku==0.3.1
django-storages==1.9.1
django==3.0.6
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9 ; python_version != '3.4'
whitenoise==5.0.1
